Refactor UnsatisfiedScopesAlert component.
